feat(compiler): make online compiler optional (#43)
OpenCL C compiler is Clang-based, which makes final LibreCL package over 300 mb in size. That is a lot for final production and deployment, especially in embedded targets, like mobile phones, or in server environment, where one controls both HW and SW, and kernels can be easily pre-compiled. This patch removes the need for an online compiler, making it possible to pack only `liblcl_icd_runtime.so` (or its OS-native counterpart) with the application, dramatically reducing app size. Since the compiler won't be available, only binary programs will be accepted.